+++ app/models/user.rb
@@ -14,6 +14,7 @@ class User < ActiveRecord::Base
   has_many :projects, -> { where deleted: false }
   has_many :categories, -> { where deleted: false }
   has_many :documents, -> { where deleted: false }
+  has_many :embeds, -> { where deleted: false }
 
   # User Methods